rename: apply file-structure
